\docType{data}
\name{Constants}
\alias{Constants}
\alias{TV_DAY}
\alias{TV_HOUR}
\alias{TV_MINUTE}
\title{Constants}
\format{Numeric constants}
\usage{
  TV_MINUTE

  TV_HOUR

  TV_DAY
}
\description{
  Constants to use with \code{\link{TimedVector}} for
  specifying time between events.
}
\details{
  \code{TV_MINUTE} A minute

  \code{TV_HOUR} An hour

  \code{TV_DAY} A day
}
\author{
  Chloe Bracis
}
\seealso{
  \code{\link{TimedVector}}
}
\keyword{datasets}

