\name{verifyTimedVector}
\alias{verifyTimedVector}
\title{Verify TimedVector}
\usage{
  verifyTimedVector(tv)
}
\arguments{
  \item{tv}{Potential \code{TimedVector} object}
}
\value{
  \code{TRUE}, if the object is a \code{TimedVector}
  \code{FALSE}, otherwise
}
\description{
  Verifies object really is a \code{\link{TimedVector}}
  (stronger checks than \code{\link{isTimedVector}}).
}
\author{
  Chloe Bracis
}
\seealso{
  \code{\link{isTimedVector}}, \code{\link{TimedVector}}
}

